+++ client/game/common/Header/FetchGameDataButton.vue
@@ -2,8 +2,8 @@
     /*******************************************************************************************************************
      * Button that fetches game data
      ******************************************************************************************************************/
-    import Icon from "../Icon/Icon.vue";
-    import Spinner from "../Spinner/Spinner.vue";
+    import Icon from "Game/common/Icon/Icon.vue";
+    import Spinner from "Game/common/Spinner/Spinner.vue";
     export default {
         data() {
             return {
@@ -22,7 +22,7 @@
         },
         computed: {
             fetching: function () {
-                return this.$store.getters.GET_fetchingGameDataFromApi;
+                return this.$store.getters.fetchingGameDataFromApi;
             }
         },
         methods: {